gates(permission and roles)

+++ app/Http/Controllers/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Idea;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
@@ -50,6 +51,8 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('user.edit', $user->id);
+
         $ideas = $user->ideas()->paginate(3);
         return view('users.edit', compact('user', 'ideas'));
     }
@@ -59,6 +62,8 @@ class UserController extends Controller
      */
     public function update(User $user)
     {
+        $this->authorize('user.edit', $user->id);
+
         $validated = request()->validate([
             'bio' => 'nullable|max:255',
             'image' => 'image',
@@ -91,6 +96,8 @@ class UserController extends Controller
     // !  Remove the specified resource from storage.
     function destroy(User $user)
     {
+        $this->authorize('user.edit', $user->id);
+
         Storage::disk('public')->delete($user->image ?? '');
         $user->update(['image' => null]);
         return redirect()->route('users.show', $user->id)->with('success', 'Image successfully deleted!');
